add IParagraph.ReplaceText(string oldValue, string newValue)
